Reduce main memory size to 1K
The memory is specified at 256MB, but the demo firmware is a little
less than 512 bytes in size. If we add a 512-byte stack, we only
need 1K of *actual* memory.

Let's only use 7 bits ([6:0]) of the 25-bit address lines, and wrap
around if accesses outside 0x8000_0000..0x8000_0400 should ever occur.